% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimisAggregate.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregate Data by ID}
\usage{
aggregateData(data, id = NULL, fun = "mean", save.path = TRUE,
  timestep = 10)
}
\arguments{
\item{data}{(data.frame) a data.frame of values to aggregate}

\item{id}{(character) the coloumn name to aggregate by}

\item{fun}{(character) the function to use in aggregation (i.e. mean)}

\item{save.path}{(character) where should a csv be written to?}

\item{timestep}{(numeric) number of observations to aggregate}
}
\value{
a data.frame of aggregated values by id
}
\description{
Aggregate Data by ID
}
