% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{daymet_flux}
\alias{daymet_flux}
\title{2015 Fluxnet stations subset to the Daymet Domain}
\format{a \code{dataframe} instance, 1 row per station with columns:
\itemize{
\item 'siteid':      A \code{character} Internal database identifier for each site
\item 'fluxnetid':    A \code{character}  Site ID within the Fluxnet network
\item 'keyid':       A \code{character}     Unknown definition
\item 'sitename':       A \code{character}      Human readable site name
\item 'country':     A \code{character}    Country location of site
\item 'land_unit':       A \code{numeric}       Region location of site
\item 'status': A \code{numeric}  Active/inactive status of site as of 2016-03-01
\item 'latitude': A \code{numeric}  Latitude coordinate
\item 'longitude':      A \code{character} Longitude coordinate
\item 'year_began':    A \code{character}  Year the site started collecting data within the FLUXNET framework
\item 'network1':       A \code{character}    Current site network
\item 'network2':     A \code{character}    Previous site network
\item 'network3':       A \code{numeric}       Other previous site network
\item 'koeppen_cl': A \code{numeric}  Koppen-Geiger climate classification
\item 'igbp_land_': A \code{numeric}  Site land cover (MODIS IGBP classification)
\item 'umd_land_c':      A \code{character}Site land cover (MODIS UMD classification)
\item 'lai_fpar':    A \code{character}  Site land cover (MODIS LAI/fpar classification)
\item 'npp_land_c':       A \code{character}      Site land cover (MODIS NPP classification)
\item 'plant_func':     A \code{character}    Unknown definition
\item 'TileID':       A \code{numeric}      Underlying DAYMET tile ID
}}
\source{
Merged NCAR Airport Dataset with DAYMET grid
}
\usage{
daymet_flux
}
\description{
Dataset containing information about USGS stations in the United States
}
\examples{
\dontrun{
 load("daymet_flux.rda")
}
}
\keyword{datasets}
