% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractXLSX_modis.R
\name{extractXLSX_modis}
\alias{extractXLSX_modis}
\title{Extract data from MODIS excel files}
\usage{
extractXLSX_modis(dir = NULL, column = "MAX")
}
\arguments{
\item{dir}{(character) the path to the XLSX files}

\item{column}{(character) what column/statistic do you want to extract?}
}
\value{
a data.frame
}
\description{
Extract data from MODIS excel files
}
