% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterStations.R
\name{filterStations}
\alias{filterStations}
\title{Filter statistics by crtieria}
\usage{
filterStations(data, stats.data, statistic, criteria, meet = TRUE)
}
\arguments{
\item{data}{a data.frame - or list of data.frames - for stations obtained with \code{getStationdata()}}

\item{stats.data}{a data.frame of statistics returned from \code{lmStats()}}

\item{statistic}{what statistic are you filtering by? Must match column name in 'stats.data'}

\item{criteria}{what is the threshold of interest?}

\item{meet}{(logical) if TRUE returns all station above criteria, else returns all stations below criteria}
}
\value{
a data.frame of stations
}
\description{
Filter statistics by crtieria
}
