% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLCc.R
\name{getLCc}
\alias{getLCc}
\title{Get Land Cover Coeffieicentes from MODIS data}
\usage{
getLCc(PET = NULL, AET = NULL, timestep = "raw")
}
\arguments{
\item{PET}{(data.frame) A data frame of PET values}

\item{AET}{(data.frame) A data frame of AET values}

\item{timestep}{(character) timestep to aggregate to: option include 'raw' (default), 'month', 'all'}
}
\value{
dataframe
}
\description{
Get Land Cover Coeffieicentes from MODIS data
}
