% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmStats.R
\name{lmStats}
\alias{lmStats}
\title{Get Regression statisitcs for a station}
\usage{
lmStats(y, x, stationData = NULL)
}
\arguments{
\item{y}{a column name to serve as the dependent variable}

\item{x}{a column name to serve as the independent variable}

\item{stationData}{a data.frame - or list of data.frames - for stations obtained with \code{getStationdata()}}
}
\value{
a data.frame of statistics for each station and test, including slope coefficenet, R2, and p-value
}
\description{
Get Regression statisitcs for a station
}
