% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberOfDays.R
\name{numberOfDays}
\alias{numberOfDays}
\title{Number of Days in a Month}
\usage{
numberOfDays(date)
}
\arguments{
\item{date}{(date) a single or vector of dates}
}
\value{
the number of days in the month of each date
}
\description{
Use this function to determine the number of days is a month
}
