% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plot Data from getStationData}
\usage{
plotData(data = NULL, stats = NULL, save.path = NULL,
  clear.folder = TRUE)
}
\arguments{
\item{data}{a data.frame - or list of data.frames - obtained with \code{getStationData}}

\item{stats}{a data.frame returned from lmStats (optional)}

\item{save.path}{a folder to save images to (optional)}

\item{clear.folder}{(logical) if saving images should the folder be cleared before running?}
}
\value{
images
}
\description{
Plot Data from getStationData
}
