% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLCc.R
\name{plotLCc}
\alias{plotLCc}
\title{Plot LCc Values}
\usage{
plotLCc(LC = NULL)
}
\arguments{
\item{LC}{(data.frame) plot LC values from getLCc}
}
\value{
a plot of LCc value time series
}
\description{
Plot LCc Values
}
