% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_carrier.R
\name{tm_carrier}
\alias{tm_carrier}
\title{TextMagic - Single Carrier Lookup}
\usage{
tm_carrier(phone = NA, country = NA, tm_user = Sys.getenv("TM_USER"),
  tm_key = Sys.getenv("TM_KEY"))
}
\arguments{
\item{phone}{character The phone number.  Required.}

\item{country}{character The phone number.  Optional.}

\item{tm_user}{character The TextMagic API username. Looks for environment variable TM_USER.  Required..}

\item{tm_key}{character The TextMagic API key. Looks for environment variable TM_KEY. Required.}
}
\value{
a list containing the response from the API
}
\description{
Retrieve the carrier for a single phone number, with country
optionally provided.
}
\details{
This function uses the TextMagic carrier lookup API.
https://www.textmagic.com/docs/api/carrier-lookup/
}
\examples{
tm_carrier(phone='6178675309', country='US')

}
\keyword{textmagic}
