% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWY.R
\name{toWY}
\alias{toWY}
\title{Converts data from Calander year to water year}
\usage{
toWY(data)
}
\arguments{
\item{data}{an input matrix where rows are years and columns are months}
}
\description{
Converts data from Calander year to water year
}
