% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio_lookup.R
\name{twilio_lookup}
\alias{twilio_lookup}
\title{Twilio - Phone Number lookup}
\usage{
twilio_lookup(phone = NA, api_sid = Sys.getenv("TWILIO_SID"),
  api_token = Sys.getenv("TWILIO_TOKEN"))
}
\arguments{
\item{phone}{character The phone number.  Required.}

\item{twilio_sid}{character The Twilio API SID. Looks for environment variable TWILIO_SID  Required..}

\item{twilio_token}{character The Twilio API token. Looks for environment variable TWILIO_TOKEN Required.}
}
\value{
a list containing the response from the API
}
\description{
Validate the phone number and return country codes and valid phone format.
}
\details{
This function uses the Twilio phone lookup API.
https://www.twilio.com/docs/lookup/api
}
\examples{
twilio_lookup(phone='6178675309')

}
\keyword{twilio}
