% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWY.R
\name{toWY}
\alias{toWY}
\title{To Water Year}
\usage{
toWY(data)
}
\arguments{
\item{data}{data organized with rows of years and columns of monthly values}
}
\description{
Convert matrix from calander year to water year
}
